poc-messaging-service-commands-main/pocMessagingServiceCommands/requirements.txt
Django==4.2.4
django_environ==0.10.0
djangorestframework==3.14.0

Requests==2.31.0

PyJWT==2.8.0
cryptography==41.0.3

confluent-kafka==2.2.0
configparser==6.0.0

uvicorn==0.23.2

django-cors-headers==4.2.0